Move HostModule(Instance) to runtime
