+++ wabt/src/main/java/com/dylibso/chicory/wabt/Wast2Json.java
@@ -6,8 +6,8 @@ import static java.nio.file.Files.createDirectories;
 import com.dylibso.chicory.log.Logger;
 import com.dylibso.chicory.log.SystemLogger;
 import com.dylibso.chicory.runtime.HostImports;
+import com.dylibso.chicory.runtime.HostModule;
 import com.dylibso.chicory.runtime.Instance;
-import com.dylibso.chicory.wasi.HostModule;
 import com.dylibso.chicory.wasi.WasiOptions;
 import com.dylibso.chicory.wasi.WasiPreview1;
 import com.dylibso.chicory.wasm.Module;
